convert random NULL usage to nullptr
Since NULL requires headers to be included to define it, and some of
our files are missing those includes depending on optimization settings,
switch over to the language nullptr builtin to avoid.

This doesn't cover all the files in the tree, just some random Linux
and common ones that I ran into build failures with.

Change-Id: I0a68538fa138148c2a0c1ec5d6dd558ffd4629b2
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/6173184
Reviewed-by: Lei Zhang <thestig@chromium.org>